format prec

+++ projects/slope-adc-stm32f411ceu/src/loop1.c
@@ -724,7 +724,7 @@ void app_loop4 ( app_t *app   )
   unsigned row = 0;
 
   // 22V range.
-  float target_[] = {
+  double target_[] = {
     11, 10.5, 10, 9.5, 9, 8.5, 8, 7.5, 7, 6.5, 6, 5.5, 5, 4.5, 4, 3.5, 3, 2.5, 2, 1.5, 1, 0.5, 0,
     -0.5, -1, -1.5, -2, -2.5, -3, -3.5, -4, -4.5, -5, -5.5, -6, -6.5, -7, -7.5, -8, -8.5, -9, -9.5, -10, -10.5, -11 } ;
 
@@ -741,7 +741,7 @@ void app_loop4 ( app_t *app   )
     double target = target_[i];
 
     // change to voltage
-    printf("voltage set %lf\n", target );
+    printf("voltage set %.1f\n", target );
     app_voltage_source_set( app, target );
 
 #if 1
